Trac #14521: Don't use NSInteger, which isn't supported on OS X 10.4

+++ src/mac-app/PreferencePanelController.m
@@ -119,7 +119,7 @@
 }
 
 
-- (void)alertDidEnd:(NSAlert *)alert returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo {
+- (void)alertDidEnd:(NSAlert *)alert returnCode:(long)returnCode contextInfo:(void *)contextInfo {
 
     if (returnCode == NSAlertFirstButtonReturn) {
         // ~/bin
